requirements.txt
aiohttp>=3.8.0
google-generativeai>=0.3.0
anthropic>=0.7.0
openai>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
